refactor: move parser tests into own file
